+++ flutter-packages/mobile/lib/presentations/verified/verified_screen.dart
@@ -91,19 +91,19 @@ class VerifiedScreen extends GetWidget<VerifiedController> {
             controller.height.value = info.height;
             controller.weight.value = info.weight;
 
-            final store = SecureMedicalStore();
-            await store.saveFromMedical(controller.userId.value, info);
+            // final store = SecureMedicalStore();
+            // await store.saveFromMedical(controller.userId.value, info);
 
-            final hasBmi = await store.s.containsKey(
-              key: 'medical_bmi ${controller.userId.value}',
-              aOptions: SecurePassportStore.aOpts,
-              iOptions: SecurePassportStore.iOpts,
-            );
-            final all = await store.s.readAll(
-              aOptions: SecurePassportStore.aOpts,
-              iOptions: SecurePassportStore.iOpts,
-            );
-            logger.d('contains birth=$hasBmi, all=$all');
+            // final hasBmi = await store.s.containsKey(
+            //   key: 'medical_bmi ${controller.userId.value}',
+            //   aOptions: SecurePassportStore.aOpts,
+            //   iOptions: SecurePassportStore.iOpts,
+            // );
+            // final all = await store.s.readAll(
+            //   aOptions: SecurePassportStore.aOpts,
+            //   iOptions: SecurePassportStore.iOpts,
+            // );
+            // logger.d('contains birth=$hasBmi, all=$all');
 
             controller.medicalNext();
           },
